requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
executing==2.2.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.8
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.3
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
